requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
gunicorn>=21.2.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
pgvector>=0.2.4
pillow>=10.0.0
numpy>=1.26.0
opencv-python>=4.8.0
insightface>=0.7.0
deepface>=0.0.79
tf-keras>=2.15.0
onnxruntime>=1.16.0
python-multipart>=0.0.6
pydantic>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.8.0
